backend/requirements.txt
pymongo
fastapi
python-dotenv
langchain
langchain_aws
pytest 
httpx 
pytest-mock